Don't use np.float128 in test_vnorm.py

+++ ufunclab/tests/test_vnorm.py
@@ -89,7 +89,7 @@ def test_axis_z(dtype):
     assert_allclose(nrm, [2, np.sqrt(30), 3, np.sqrt(204)])
 
 
-@pytest.mark.parametrize('dtype', [np.float64, np.float128,
+@pytest.mark.parametrize('dtype', [np.float32, np.float64,
                                    np.complex64, np.complex128])
 def test_nontrivial_strides(dtype):
     a = np.zeros((3, 8), dtype=dtype)
